Move _TEST_PRESET_PATCHES to a pytest-independent module
roborev caught the previous fix: importing _patch_presets_for_testing
from tests.conftest pulled in conftest's top-level `import pytest`,
which is only in the dev extra. tests/generate_intermediates.py is
documented to run standalone (`uv run python tests/generate_intermediates.py`),
so a clean environment without the dev extra would fail at import
before ever reaching argparse.

Extracted the table and the patching function into tests/preset_patches.py,
which imports only yaml/pathlib. Both tests/conftest.py and
tests/generate_intermediates.py now import from there - one copy of the
IPCA regularization fix, reachable without pytest.

Verified: `uv run --no-dev python tests/generate_intermediates.py --help`
runs clean (the exact case roborev flagged), the module import path
still works under pytest, and 11b_ipca still passes locally.
